Rename hashupdater file -> loginmanager
