+++ cmake/SourceFileList.cmake
@@ -87,10 +87,10 @@ set(eoserv_ALL_SOURCE_FILES
 	src/handlers/handlers.hpp
 	src/hash.cpp
 	src/hash.hpp
-	src/hashupdater.cpp
-	src/hashupdater.hpp
 	src/i18n.cpp
 	src/i18n.hpp
+	src/loginmanager.cpp
+	src/loginmanager.hpp
 	src/map.cpp
 	src/map.hpp
 	src/nanohttp.cpp
