+++ src/loginmanager.hpp
